Rebase envoy and use sha-1 cluster names (#613)
* update proxy sha

* update goldens

* gazelle

* label tests as small

* buildifier
