+++ client/proxy/BUILD
@@ -12,6 +12,7 @@ go_library(
 
 go_test(
     name = "go_default_test",
+    size = "small",
     srcs = ["proxy_test.go"],
     library = ":go_default_library",
     deps = [
