+++ cmd/istioctl/BUILD
@@ -3,10 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "collateral.go",
         "inject.go",
         "main.go",
         "mixer.go",
-        "collateral.go",
     ],
     visibility = ["//visibility:private"],
     deps = [
@@ -19,7 +19,6 @@ go_library(
         "//platform/kube/inject:go_default_library",
         "@com_github_ghodss_yaml//:go_default_library",
         "@com_github_golang_glog//:go_default_library",
-        "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_googleapis_googleapis//:go_default_library",
         "@com_github_hashicorp_go_multierror//:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
