+++ proxy/envoy/testdata/cds-egress.json.golden
@@ -1,19 +1,7 @@
 {
   "clusters": [
    {
-    "name": "out.httpbin.default.svc.cluster.local|http",
-    "service_name": "httpbin.default.svc.cluster.local|http",
-    "connect_timeout_ms": 1000,
-    "type": "strict_dns",
-    "lb_type": "round_robin",
-    "hosts": [
-     {
-      "url": "tcp://httpbin.org:80"
-     }
-    ]
-   },
-   {
-    "name": "out.httpsbin.default.svc.cluster.local|https",
+    "name": "out.242bc3028e0f3fe0682e6d972e167ab415b2321d",
     "service_name": "httpsbin.default.svc.cluster.local|https",
     "connect_timeout_ms": 1000,
     "type": "strict_dns",
@@ -24,6 +12,18 @@
      }
     ],
     "ssl_context": {}
+   },
+   {
+    "name": "out.ae8d3361601f8293abe6ac5e4d807124612cf42e",
+    "service_name": "httpbin.default.svc.cluster.local|http",
+    "connect_timeout_ms": 1000,
+    "type": "strict_dns",
+    "lb_type": "round_robin",
+    "hosts": [
+     {
+      "url": "tcp://httpbin.org:80"
+     }
+    ]
    }
   ]
  }
