+++ proxy/envoy/testdata/cds-ssl-context.json.golden
@@ -1,11 +1,15 @@
 {
   "clusters": [
    {
-    "name": "out.hello.default.svc.cluster.local|http",
-    "service_name": "hello.default.svc.cluster.local|http",
+    "name": "out.242bc3028e0f3fe0682e6d972e167ab415b2321d",
     "connect_timeout_ms": 1000,
-    "type": "sds",
+    "type": "strict_dns",
     "lb_type": "round_robin",
+    "hosts": [
+     {
+      "url": "tcp://istio-egress:80"
+     }
+    ],
     "ssl_context": {
      "cert_chain_file": "/etc/certs/cert-chain.pem",
      "private_key_file": "/etc/certs/key.pem",
@@ -14,7 +18,7 @@
     }
    },
    {
-    "name": "out.hello.default.svc.cluster.local|http-status",
+    "name": "out.81c187d71467b1608736c57bb0734f9ef9b68f7d",
     "service_name": "hello.default.svc.cluster.local|http-status",
     "connect_timeout_ms": 1000,
     "type": "sds",
@@ -27,41 +31,7 @@
     }
    },
    {
-    "name": "out.httpbin.default.svc.cluster.local|http",
-    "connect_timeout_ms": 1000,
-    "type": "strict_dns",
-    "lb_type": "round_robin",
-    "hosts": [
-     {
-      "url": "tcp://istio-egress:80"
-     }
-    ],
-    "ssl_context": {
-     "cert_chain_file": "/etc/certs/cert-chain.pem",
-     "private_key_file": "/etc/certs/key.pem",
-     "ca_cert_file": "/etc/certs/root-cert.pem",
-     "verify_subject_alt_name": []
-    }
-   },
-   {
-    "name": "out.httpsbin.default.svc.cluster.local|https",
-    "connect_timeout_ms": 1000,
-    "type": "strict_dns",
-    "lb_type": "round_robin",
-    "hosts": [
-     {
-      "url": "tcp://istio-egress:80"
-     }
-    ],
-    "ssl_context": {
-     "cert_chain_file": "/etc/certs/cert-chain.pem",
-     "private_key_file": "/etc/certs/key.pem",
-     "ca_cert_file": "/etc/certs/root-cert.pem",
-     "verify_subject_alt_name": []
-    }
-   },
-   {
-    "name": "out.world.default.svc.cluster.local|http",
+    "name": "out.a2263b90cd24e500e9ed95f79ae47eabdc77dc74",
     "service_name": "world.default.svc.cluster.local|http",
     "connect_timeout_ms": 1000,
     "type": "sds",
@@ -77,7 +47,24 @@
     }
    },
    {
-    "name": "out.world.default.svc.cluster.local|http-status",
+    "name": "out.ae8d3361601f8293abe6ac5e4d807124612cf42e",
+    "connect_timeout_ms": 1000,
+    "type": "strict_dns",
+    "lb_type": "round_robin",
+    "hosts": [
+     {
+      "url": "tcp://istio-egress:80"
+     }
+    ],
+    "ssl_context": {
+     "cert_chain_file": "/etc/certs/cert-chain.pem",
+     "private_key_file": "/etc/certs/key.pem",
+     "ca_cert_file": "/etc/certs/root-cert.pem",
+     "verify_subject_alt_name": []
+    }
+   },
+   {
+    "name": "out.bde94496eb59ec2ed5b81392a1d32377960660b8",
     "service_name": "world.default.svc.cluster.local|http-status",
     "connect_timeout_ms": 1000,
     "type": "sds",
@@ -90,6 +77,19 @@
       "istio:serviceaccount1",
       "istio:serviceaccount2"
      ]
+    }
+   },
+   {
+    "name": "out.e5c9564b7c4dbb0355a4f740e9d29277ccca97cd",
+    "service_name": "hello.default.svc.cluster.local|http",
+    "connect_timeout_ms": 1000,
+    "type": "sds",
+    "lb_type": "round_robin",
+    "ssl_context": {
+     "cert_chain_file": "/etc/certs/cert-chain.pem",
+     "private_key_file": "/etc/certs/key.pem",
+     "ca_cert_file": "/etc/certs/root-cert.pem",
+     "verify_subject_alt_name": []
     }
    }
   ]
