+++ proxy/envoy/testdata/envoy-egress-auth.json.golden
