+++ proxy/envoy/testdata/envoy-egress.json.golden
