+++ proxy/envoy/testdata/envoy-v1-auth.json.golden
@@ -152,7 +152,7 @@
             "route_config": {
               "routes": [
                 {
-                  "cluster": "out.hello.default.svc.cluster.local|custom",
+                  "cluster": "out.de6d66d4dd5f542e5f61882eb466189eb68ebe88",
                   "destination_ip_list": [
                     "10.1.0.0/32"
                   ]
@@ -374,7 +374,7 @@
             "route_config": {
               "routes": [
                 {
-                  "cluster": "out.world.default.svc.cluster.local|custom",
+                  "cluster": "out.5898aa4379cc19c8f1bb3b7915ee8e0e32ddc6a6",
                   "destination_ip_list": [
                     "10.2.0.0/32"
                   ]
@@ -444,15 +444,15 @@
         ]
       },
       {
-        "name": "out.hello.default.svc.cluster.local|custom",
-        "service_name": "hello.default.svc.cluster.local|custom",
+        "name": "out.5898aa4379cc19c8f1bb3b7915ee8e0e32ddc6a6",
+        "service_name": "world.default.svc.cluster.local|custom",
         "connect_timeout_ms": 1000,
         "type": "sds",
         "lb_type": "round_robin"
       },
       {
-        "name": "out.world.default.svc.cluster.local|custom",
-        "service_name": "world.default.svc.cluster.local|custom",
+        "name": "out.de6d66d4dd5f542e5f61882eb466189eb68ebe88",
+        "service_name": "hello.default.svc.cluster.local|custom",
         "connect_timeout_ms": 1000,
         "type": "sds",
         "lb_type": "round_robin"
