+++ proxy/envoy/testdata/rds-egress.json.golden
@@ -8,7 +8,7 @@
     "routes": [
      {
       "prefix": "/",
-      "cluster": "out.httpbin.default.svc.cluster.local|http",
+      "cluster": "out.ae8d3361601f8293abe6ac5e4d807124612cf42e",
       "auto_host_rewrite": true
      }
     ]
@@ -21,7 +21,7 @@
     "routes": [
      {
       "prefix": "/",
-      "cluster": "out.httpsbin.default.svc.cluster.local|https",
+      "cluster": "out.242bc3028e0f3fe0682e6d972e167ab415b2321d",
       "auto_host_rewrite": true
      }
     ]
