+++ proxy/envoy/testdata/rds-ingress-ssl.json.golden
@@ -9,7 +9,7 @@
      {
       "prefix": "/bar",
       "host_rewrite": "foo.default.svc.cluster.local:80",
-      "cluster": "out.foo.default.svc.cluster.local|http-status"
+      "cluster": "out.ce697a7c2c3c535648c2bf2c3f7b9fe599e8f5f4"
      }
     ]
    }
