+++ proxy/envoy/testdata/rds-v0.json.golden
@@ -19,7 +19,7 @@
     "routes": [
      {
       "prefix": "/",
-      "cluster": "out.hello.default.svc.cluster.local|http"
+      "cluster": "out.e5c9564b7c4dbb0355a4f740e9d29277ccca97cd"
      }
     ]
    },
@@ -41,7 +41,7 @@
      {
       "prefix": "/",
       "host_rewrite": "httpbin.default.svc.cluster.local",
-      "cluster": "out.httpbin.default.svc.cluster.local|http"
+      "cluster": "out.ae8d3361601f8293abe6ac5e4d807124612cf42e"
      }
     ]
    },
@@ -64,7 +64,7 @@
     "routes": [
      {
       "prefix": "/",
-      "cluster": "out.world.default.svc.cluster.local|http"
+      "cluster": "out.a2263b90cd24e500e9ed95f79ae47eabdc77dc74"
      }
     ]
    }
