+++ proxy/envoy/testdata/rds-weighted.json.golden
@@ -19,7 +19,7 @@
     "routes": [
      {
       "prefix": "/",
-      "cluster": "out.hello.default.svc.cluster.local|http"
+      "cluster": "out.e5c9564b7c4dbb0355a4f740e9d29277ccca97cd"
      }
     ]
    },
@@ -41,7 +41,7 @@
      {
       "prefix": "/",
       "host_rewrite": "httpbin.default.svc.cluster.local",
-      "cluster": "out.httpbin.default.svc.cluster.local|http"
+      "cluster": "out.ae8d3361601f8293abe6ac5e4d807124612cf42e"
      }
     ]
    },
@@ -68,11 +68,11 @@
       "weighted_clusters": {
        "clusters": [
         {
-         "name": "out.world.default.svc.cluster.local|http|version=v0",
+         "name": "out.c76febe0f151b2f8abe0f377d2052c0fbbfb959d",
          "weight": 75
         },
         {
-         "name": "out.world.default.svc.cluster.local|http|version=v1",
+         "name": "out.66fcc955b8875b19844f9eaf6cfda47c778c609e",
          "weight": 25
         }
        ]
